backend/requirements.txt
fastapi>=0.116.1,<0.117.0
uvicorn>=0.35.0,<0.36.0
python-multipart>=0.0.20,<0.0.21
pydantic>=2.0.0,<3.0.0
pandas>=2.2.0,<2.3.0
openpyxl>=3.1.2,<3.2.0
aiofiles>=23.2.1,<24.0.0
openai>=1.0.0,<2.0.0
pyjwt>=2.8.0,<3.0.0
email-validator>=2.0.0

# Cloud Database & Storage
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-dotenv>=1.0.0

# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# Alternative Cloud Services (optional)
supabase>=2.3.0
cloudinary>=1.36.0